[CHORE] packaging refactoring
