+++ app/src/main/java/com/example/sobok_android/data/api/LoginService.kt
@@ -1,12 +1,11 @@
 package com.example.sobok_android.data.api
 
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.request.ReqSignUpSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignUpSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignUpSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignInSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignUpSuccessData
 import retrofit2.http.Body
 import retrofit2.http.POST
-import retrofit2.http.Query
 
 interface LoginService {
     //POST SignIn
