+++ app/src/main/java/com/example/sobok_android/data/datasource/login/SignInDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.sobok_android.data.datasource.login
 
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignInSuccessData
 
 interface SignInDataSource {
