+++ app/src/main/java/com/example/sobok_android/data/datasource/login/SignInRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.example.sobok_android.data.datasource.login
 
 import com.example.sobok_android.data.api.LoginService
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignInSuccessData
 
 class SignInRemoteDataSource(private val loginService: LoginService) : SignInDataSource {
     override suspend fun postSignIn(body: ReqSignInSuccessData): ResSignInSuccessData {
