+++ app/src/main/java/com/example/sobok_android/data/datasource/login/SignUpDataSource.kt
@@ -1,9 +1,7 @@
 package com.example.sobok_android.data.datasource.login
 
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.request.ReqSignUpSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignUpSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignUpSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignUpSuccessData
 
 interface SignUpDataSource {
     suspend fun postSignUp(body: ReqSignUpSuccessData): ResSignUpSuccessData
