+++ app/src/main/java/com/example/sobok_android/data/datasource/login/SignUpRemoteDataSource.kt
@@ -1,10 +1,8 @@
 package com.example.sobok_android.data.datasource.login
 
 import com.example.sobok_android.data.api.LoginService
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.request.ReqSignUpSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignUpSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignUpSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignUpSuccessData
 
 class SignUpRemoteDataSource(private val loginService: LoginService) : SignUpDataSource {
     override suspend fun postSignUp(body: ReqSignUpSuccessData): ResSignUpSuccessData {
