+++ app/src/main/java/com/example/sobok_android/data/mapper/SignInMapper.kt
@@ -1,6 +1,6 @@
 package com.example.sobok_android.data.mapper
 
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignInSuccessData
 import com.example.sobok_android.domain.model.login.request.SignInSuccessData
 
 object SignInMapper {
