+++ app/src/main/java/com/example/sobok_android/data/mapper/SignUpMapper.kt
@@ -1,8 +1,6 @@
 package com.example.sobok_android.data.mapper
 
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
-import com.example.sobok_android.data.model.response.login.request.ResSignUpSuccessData
-import com.example.sobok_android.domain.model.login.request.SignInSuccessData
+import com.example.sobok_android.data.model.response.login.ResSignUpSuccessData
 import com.example.sobok_android.domain.model.login.request.SignUpSuccessData
 
 object SignUpMapper {
