+++ app/src/main/java/com/example/sobok_android/data/model/request/login/ReqSignInSuccessData.kt
@@ -1,4 +1,4 @@
-package com.example.sobok_android.data.model.request
+package com.example.sobok_android.data.model.request.login
 
 data class ReqSignInSuccessData(
     val email : String,
