+++ app/src/main/java/com/example/sobok_android/data/repository/login/request/SignInRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.example.sobok_android.data.repository.login.request
 
 import com.example.sobok_android.data.datasource.login.SignInDataSource
 import com.example.sobok_android.data.mapper.SignInMapper
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
 import com.example.sobok_android.domain.model.login.request.SignInSuccessData
 import com.example.sobok_android.domain.repository.login.request.SignInRepository
