+++ app/src/main/java/com/example/sobok_android/domain/model/login/request/SignInSuccessData.kt
@@ -1,6 +1,5 @@
 package com.example.sobok_android.domain.model.login.request
 
-import com.example.sobok_android.data.model.response.login.request.ResSignInSuccessData
 import com.google.gson.annotations.SerializedName
 
 data class SignInSuccessData(
