+++ app/src/main/java/com/example/sobok_android/domain/repository/login/request/SignInRepository.kt
@@ -1,6 +1,6 @@
 package com.example.sobok_android.domain.repository.login.request
 
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
 import com.example.sobok_android.domain.model.login.request.SignInSuccessData
 
 interface SignInRepository {
