+++ app/src/main/java/com/example/sobok_android/presentation/view/user/request/viewmodel/SignInViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignInSuccessData
 import com.example.sobok_android.data.sharedpref.SobokSharedPreference
 import com.example.sobok_android.domain.model.login.request.SignInSuccessData
 import com.example.sobok_android.domain.repository.login.request.SignInRepository
