+++ app/src/main/java/com/example/sobok_android/presentation/view/user/request/viewmodel/SignUpViewModel.kt
@@ -5,11 +5,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.sobok_android.data.model.request.ReqSignInSuccessData
-import com.example.sobok_android.data.model.request.ReqSignUpSuccessData
-import com.example.sobok_android.domain.model.login.request.SignInSuccessData
+import com.example.sobok_android.data.model.request.login.ReqSignUpSuccessData
 import com.example.sobok_android.domain.model.login.request.SignUpSuccessData
-import com.example.sobok_android.domain.repository.login.request.SignInRepository
 import com.example.sobok_android.domain.repository.login.request.SignUpRepository
 import kotlinx.coroutines.launch
